app/requirements.txt
gradio==4.12.0
psycopg2-binary==2.9.9
pgvector==0.2.4
sentence-transformers==2.3.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.26.2